Adding the decryption of the email password to the config file
